separate reactions from types
