+++ include/mechanism_configuration/v1/parser_types.hpp
@@ -5,9 +5,9 @@
 #pragma once
 
 #include <mechanism_configuration/errors.hpp>
-#include <mechanism_configuration/parse_status.hpp>
+#include <mechanism_configuration/v1/models.hpp>
 #include <mechanism_configuration/v1/types.hpp>
-#include <mechanism_configuration/v1/utils.hpp>
+
 
 #include <yaml-cpp/yaml.h>
 
@@ -18,156 +18,18 @@ namespace mechanism_configuration
 {
   namespace v1
   {
-    std::pair<Errors, std::vector<v1::types::Species>> ParseSpecies(const YAML::Node& objects);
+    std::pair<Errors, std::vector<types::Species>> ParseSpecies(const YAML::Node& objects);
 
-    class CondensedPhaseArrheniusParser : public IReactionParser
-    {
-     public:
-      Errors parse(
-          const YAML::Node& object,
-          const std::vector<types::Species>& existing_species,
-          const std::vector<types::Phase>& existing_phases,
-          types::Reactions& reactions) override;
-    };
+    std::pair<Errors, std::vector<types::Phase>> ParsePhases(const YAML::Node& objects, const std::vector<v1::types::Species> existing_species);
 
-    class TroeParser : public IReactionParser
-    {
-     public:
-      Errors parse(
-          const YAML::Node& object,
-          const std::vector<types::Species>& existing_species,
-          const std::vector<types::Phase>& existing_phases,
-          types::Reactions& reactions) override;
-    };
+    std::pair<Errors, types::Models> ParseModels(const YAML::Node& objects, const std::vector<types::Phase>& existing_phases);
 
-    class BranchedParser : public IReactionParser
-    {
-     public:
-      Errors parse(
-          const YAML::Node& object,
-          const std::vector<types::Species>& existing_species,
-          const std::vector<types::Phase>& existing_phases,
-          types::Reactions& reactions) override;
-    };
+    std::pair<Errors, types::ReactionComponent> ParseReactionComponent(const YAML::Node& object);
 
-    class TunnelingParser : public IReactionParser
-    {
-     public:
-      Errors parse(
-          const YAML::Node& object,
-          const std::vector<types::Species>& existing_species,
-          const std::vector<types::Phase>& existing_phases,
-          types::Reactions& reactions) override;
-    };
+    std::pair<Errors, std::vector<types::ReactionComponent>> ParseReactantsOrProducts(const std::string& key, const YAML::Node& object);
 
-    class SurfaceParser : public IReactionParser
-    {
-     public:
-      Errors parse(
-          const YAML::Node& object,
-          const std::vector<types::Species>& existing_species,
-          const std::vector<types::Phase>& existing_phases,
-          types::Reactions& reactions) override;
-    };
+    std::pair<Errors, types::Reactions>
+    ParseReactions(const YAML::Node& objects, const std::vector<types::Species>& existing_species, const std::vector<types::Phase>& existing_phases);
 
-    class CondensedPhasePhotolysisParser : public IReactionParser
-    {
-     public:
-      Errors parse(
-          const YAML::Node& object,
-          const std::vector<types::Species>& existing_species,
-          const std::vector<types::Phase>& existing_phases,
-          types::Reactions& reactions) override;
-    };
-
-    class SimpolPhaseTransferParser : public IReactionParser
-    {
-     public:
-      Errors parse(
-          const YAML::Node& object,
-          const std::vector<types::Species>& existing_species,
-          const std::vector<types::Phase>& existing_phases,
-          types::Reactions& reactions) override;
-    };
-
-    class EmissionParser : public IReactionParser
-    {
-     public:
-      Errors parse(
-          const YAML::Node& object,
-          const std::vector<types::Species>& existing_species,
-          const std::vector<types::Phase>& existing_phases,
-          types::Reactions& reactions) override;
-    };
-
-    class PhotolysisParser : public IReactionParser
-    {
-     public:
-      Errors parse(
-          const YAML::Node& object,
-          const std::vector<types::Species>& existing_species,
-          const std::vector<types::Phase>& existing_phases,
-          types::Reactions& reactions) override;
-    };
-
-    class FirstOrderLossParser : public IReactionParser
-    {
-     public:
-      Errors parse(
-          const YAML::Node& object,
-          const std::vector<types::Species>& existing_species,
-          const std::vector<types::Phase>& existing_phases,
-          types::Reactions& reactions) override;
-    };
-
-    class AqueousEquilibriumParser : public IReactionParser
-    {
-     public:
-      Errors parse(
-          const YAML::Node& object,
-          const std::vector<types::Species>& existing_species,
-          const std::vector<types::Phase>& existing_phases,
-          types::Reactions& reactions) override;
-    };
-
-    class WetDepositionParser : public IReactionParser
-    {
-     public:
-      Errors parse(
-          const YAML::Node& object,
-          const std::vector<types::Species>& existing_species,
-          const std::vector<types::Phase>& existing_phases,
-          types::Reactions& reactions) override;
-    };
-
-    class HenrysLawParser : public IReactionParser
-    {
-     public:
-      Errors parse(
-          const YAML::Node& object,
-          const std::vector<types::Species>& existing_species,
-          const std::vector<types::Phase>& existing_phases,
-          types::Reactions& reactions) override;
-    };
-
-    class ArrheniusParser : public IReactionParser
-    {
-     public:
-      Errors parse(
-          const YAML::Node& object,
-          const std::vector<types::Species>& existing_species,
-          const std::vector<types::Phase>& existing_phases,
-          types::Reactions& reactions) override;
-    };
-
-    class UserDefinedParser : public IReactionParser
-    {
-     public:
-      Errors parse(
-          const YAML::Node& object,
-          const std::vector<types::Species>& existing_species,
-          const std::vector<types::Phase>& existing_phases,
-          types::Reactions& reactions) override;
-    };
   }  // namespace v1
 }  // namespace mechanism_configuration
