+++ include/mechanism_configuration/v1/utils.hpp
@@ -6,6 +6,7 @@
 
 #include <mechanism_configuration/errors.hpp>
 #include <mechanism_configuration/parse_status.hpp>
+#include <mechanism_configuration/v1/models.hpp>
 #include <mechanism_configuration/v1/types.hpp>
 #include <mechanism_configuration/v1/validation.hpp>
 
@@ -20,29 +21,8 @@ namespace mechanism_configuration
 {
   namespace v1
   {
-    class IReactionParser
-    {
-     public:
-      virtual Errors parse(
-          const YAML::Node& object,
-          const std::vector<v1::types::Species>& existing_species,
-          const std::vector<v1::types::Phase>& existing_phases,
-          v1::types::Reactions& reactions) = 0;
-      virtual ~IReactionParser() = default;
-    };
 
     std::unordered_map<std::string, std::string> GetComments(const YAML::Node& object);
-
-    std::pair<Errors, std::vector<v1::types::Phase>> ParsePhases(const YAML::Node& objects, const std::vector<v1::types::Species> existing_species);
-
-    std::pair<Errors, v1::types::ReactionComponent> ParseReactionComponent(const YAML::Node& object);
-
-    std::pair<Errors, std::vector<v1::types::ReactionComponent>> ParseReactantsOrProducts(const std::string& key, const YAML::Node& object);
-
-    std::pair<Errors, types::Reactions>
-    ParseReactions(const YAML::Node& objects, const std::vector<types::Species>& existing_species, const std::vector<types::Phase>& existing_phases);
-
-    std::pair<Errors, types::Models> ParseModels(const YAML::Node& objects, const std::vector<types::Phase>& existing_phases);
 
     template<typename T>
     bool ContainsUniqueObjectsByName(const std::vector<T>& collection)
